off_chain/requirements.txt
loguru==0.7.2
web3==6.11.3
openai==0.28
python-dotenv==1.0.0
